Introduce new /spaces specific `config.js`
... and reduce the complexity around `next.config.js`, `webpack.config.js` & co.
