+++ .gitignore
@@ -11,7 +11,7 @@
 # next.js
 /.next/
 /out/
-next.config.js
+config.js
 
 # production
 /build
