+++ lib/Auth.js
@@ -10,7 +10,7 @@ import MyPadsAuthProvider from './auth/MyPadsAuthProvider';
     Mapping of authentication provider types and which corresponding class should be used to actually perform
     any logic specifically related to the given authentication provider.
 
-    You can define multiple authentication providers in the `next.config.js` file.
+    You can define multiple authentication providers in the `config.js` file.
  */
 const AuthProviders = {
     matrix: MatrixAuthProvider,
