+++ lib/auth/MyPadsAuthProvider.js
@@ -17,7 +17,7 @@ class MyPadsAuthProvider {
 
     async signin(usr, pwd) {
         // successfull login revokes all previouse generated tokens for this user
-        // if no api url is defined in next.config.js we exit the function
+        // if no api url is defined in config.js we exit the function
         if (!this.url) return;
         const login = await this._login(this.url, usr, pwd);
         this.token = login.token;
